+++ src/main/java/wooteco/subway/admin/dto/req/LineRequest.java
@@ -1,8 +1,8 @@
-package wooteco.subway.admin.dto;
-
-import wooteco.subway.admin.domain.Line;
+package wooteco.subway.admin.dto.req;
 
 import java.time.LocalTime;
+
+import wooteco.subway.admin.domain.Line;
 
 public class LineRequest {
     private String name;
