+++ src/main/java/wooteco/subway/admin/dto/req/LineStationCreateRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.admin.dto;
+package wooteco.subway.admin.dto.req;
 
 public class LineStationCreateRequest {
     private Long preStationId;
