+++ src/main/java/wooteco/subway/admin/dto/req/StationCreateRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.admin.dto;
+package wooteco.subway.admin.dto.req;
 
 import wooteco.subway.admin.domain.Station;
