+++ src/main/java/wooteco/subway/admin/dto/res/LineResponse.java
@@ -1,13 +1,13 @@
-package wooteco.subway.admin.dto;
+package wooteco.subway.admin.dto.res;
 
 import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.util.List;
+import java.util.stream.Collectors;
 
 import wooteco.subway.admin.domain.Line;
-import wooteco.subway.admin.domain.Station;
 
-public class LineDetailResponse {
+public class LineResponse {
     private Long id;
     private String name;
     private LocalTime startTime;
@@ -15,14 +15,23 @@ public class LineDetailResponse {
     private int intervalTime;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
-    private List<StationResponse> stations;
 
-    public LineDetailResponse() {
+    public static LineResponse of(Line line) {
+        return new LineResponse(line.getId(), line.getName(), line.getStartTime(),
+            line.getEndTime(), line.getIntervalTime(), line.getCreatedAt(), line.getUpdatedAt());
     }
 
-    public LineDetailResponse(Long id, String name, LocalTime startTime, LocalTime endTime,
-        int intervalTime, LocalDateTime createdAt, LocalDateTime updatedAt,
-        List<Station> stations) {
+    public static List<LineResponse> listOf(List<Line> lines) {
+        return lines.stream()
+            .map(it -> LineResponse.of(it))
+            .collect(Collectors.toList());
+    }
+
+    public LineResponse() {
+    }
+
+    public LineResponse(Long id, String name, LocalTime startTime, LocalTime endTime,
+        int intervalTime, LocalDateTime createdAt, LocalDateTime updatedAt) {
         this.id = id;
         this.name = name;
         this.startTime = startTime;
@@ -30,13 +39,6 @@ public class LineDetailResponse {
         this.intervalTime = intervalTime;
         this.createdAt = createdAt;
         this.updatedAt = updatedAt;
-        this.stations = StationResponse.listOf(stations);
-    }
-
-    public static LineDetailResponse of(Line line, List<Station> stations) {
-        return new LineDetailResponse(line.getId(), line.getName(), line.getStartTime(),
-            line.getEndTime(), line.getIntervalTime(), line.getCreatedAt(), line.getUpdatedAt(),
-            stations);
     }
 
     public Long getId() {
@@ -65,9 +67,5 @@ public class LineDetailResponse {
 
     public LocalDateTime getUpdatedAt() {
         return updatedAt;
-    }
-
-    public List<StationResponse> getStations() {
-        return stations;
     }
 }
