+++ src/main/java/wooteco/subway/admin/dto/res/WholeSubwayResponse.java
@@ -1,4 +1,4 @@
-package wooteco.subway.admin.dto;
+package wooteco.subway.admin.dto.res;
 
 import java.util.List;
