+++ src/test/java/wooteco/subway/admin/acceptance/LineAcceptanceTest.java
@@ -8,8 +8,8 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import wooteco.subway.admin.dto.LineDetailResponse;
-import wooteco.subway.admin.dto.LineResponse;
+import wooteco.subway.admin.dto.res.LineDetailResponse;
+import wooteco.subway.admin.dto.res.LineResponse;
 
 public class LineAcceptanceTest extends AcceptanceTest {
     @DisplayName("지하철 노선을 관리한다")
