+++ src/test/java/wooteco/subway/admin/api/LineControllerTest.java
@@ -20,8 +20,8 @@ import wooteco.subway.admin.config.ETagHeaderFilter;
 import wooteco.subway.admin.controller.LineController;
 import wooteco.subway.admin.domain.Line;
 import wooteco.subway.admin.domain.Station;
-import wooteco.subway.admin.dto.LineDetailResponse;
-import wooteco.subway.admin.dto.WholeSubwayResponse;
+import wooteco.subway.admin.dto.res.LineDetailResponse;
+import wooteco.subway.admin.dto.res.WholeSubwayResponse;
 import wooteco.subway.admin.service.LineService;
 
 @WebMvcTest(controllers = {LineController.class})
